requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tkcalendar==1.6.1
openpyxl==3.1.2
pandas==2.1.4
